refactor(api): extract pages + global-blocks domains (#35)
